Add lombok getters, setters, constructors. Fixed sql scripts

+++ src/main/resources/liquibase/db_changelog/v1.0/create_table_candidate_feedback.sql
@@ -3,7 +3,7 @@ CREATE TABLE candidate_feedback(
     candidate_id BIGINT NOT NULL,
     hr_id BIGINT NOT NULL,
     comment VARCHAR(2000),
-    rating SMALLINT NOT NULL,
+    rating INTEGER NOT NULL,
     FOREIGN KEY (hr_id) REFERENCES hr(id),
     FOREIGN KEY (candidate_id) REFERENCES candidate(id)
 );
